+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 
 ## [Unreleased]
+
+
+## [2.4.0] - 2018-09-02
 ### Added
 - `movePopup()`: Share a picker among multiple parents.
 - **silent** flag in `setColor()` (#4).
@@ -56,7 +59,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Inline styles replaced with CSS for easier customization
 
 
-[Unreleased]: https://github.com/Sphinxxxx/vanilla-picker/compare/v2.3.0...HEAD
+[Unreleased]: https://github.com/Sphinxxxx/vanilla-picker/compare/v2.4.0...HEAD
+[2.4.0]:      https://github.com/Sphinxxxx/vanilla-picker/compare/v2.3.0...v2.4.0
 [2.3.0]:      https://github.com/Sphinxxxx/vanilla-picker/compare/v2.2.1...v2.3.0
 [2.2.1]:      https://github.com/Sphinxxxx/vanilla-picker/compare/v2.2.0...v2.2.1
 [2.2.0]:      https://github.com/Sphinxxxx/vanilla-picker/compare/v2.1.0...v2.2.0